refactor: api, query 디렉토리 변경
