+++ src/apis/types/common.ts
@@ -1,0 +1,5 @@
+export type ResponseFormat<T> = {
+  status: string;
+  data: T;
+  message: string;
+};
